Begin moving Audio code to be in-line with AziAudio

+++ Source/HLEAudio/AudioHLEProcessor.h
@@ -64,7 +64,7 @@ private:
 
 
 public:
-	ALIGNED_TYPE(u8, Buffer[0x10000], 16);	// Seems excesively large? 0x1000 should be enough, but will require to make many changes, ex update the bitfields 
+	ALIGNED_TYPE(s16, Buffer[0x10000], 16);	// Seems excesively large? 0x1000 should be enough, but will require to make many changes, ex update the bitfields 
 	u16		ADPCMTable[0x88];
 
 	//u32		Segments[16];		// 0x0320
